refactor: move AuthorizationRequest to AuthorizationCheckBehavior
This request is used to determine if a user is authorized to perform an action. It is best described by the
AuthorizationCheckBehavior, as this is the class responsible for checking the authorizations.
