+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/streamprocessor/AuthorizableCommandProcessor.java
@@ -9,8 +9,7 @@ package io.camunda.zeebe.engine.processing.streamprocessor;
 
 import io.camunda.zeebe.engine.EngineConfiguration;
 import io.camunda.zeebe.engine.processing.identity.AuthorizationCheckBehavior;
-import io.camunda.zeebe.engine.processing.streamprocessor.AuthorizableDistributionProcessor.Authorizable;
-import io.camunda.zeebe.engine.processing.streamprocessor.AuthorizableDistributionProcessor.AuthorizationRequest;
+import io.camunda.zeebe.engine.processing.identity.AuthorizationCheckBehavior.AuthorizationRequest;
 import io.camunda.zeebe.engine.processing.streamprocessor.TypedRecordProcessor.ProcessingError;
 import io.camunda.zeebe.engine.processing.streamprocessor.writers.StateWriter;
 import io.camunda.zeebe.engine.processing.streamprocessor.writers.TypedCommandWriter;
@@ -21,8 +20,6 @@ import io.camunda.zeebe.engine.state.immutable.ProcessingState;
 import io.camunda.zeebe.protocol.impl.record.UnifiedRecordValue;
 import io.camunda.zeebe.protocol.record.RejectionType;
 import io.camunda.zeebe.protocol.record.intent.Intent;
-import io.camunda.zeebe.protocol.record.value.AuthorizationResourceType;
-import io.camunda.zeebe.protocol.record.value.PermissionType;
 import io.camunda.zeebe.stream.api.records.TypedRecord;
 
 /**
@@ -60,8 +57,7 @@ public class AuthorizableCommandProcessor<T extends UnifiedRecordValue>
   public boolean onCommand(final TypedRecord<T> command, final CommandControl<T> controller) {
     final var authorizationRequest = delegate.getAuthorizationRequest();
 
-    if (authorizationCheckBehavior.isAuthorized(
-        command, authorizationRequest.resourceType(), authorizationRequest.permissionType())) {
+    if (authorizationCheckBehavior.isAuthorized(command, authorizationRequest)) {
       return delegate.onCommand(command, controller);
     } else {
       final var errorMessage =
@@ -87,9 +83,6 @@ public class AuthorizableCommandProcessor<T extends UnifiedRecordValue>
   public ProcessingError tryHandleError(final TypedRecord<T> command, final Throwable error) {
     return delegate.tryHandleError(command, error);
   }
-
-  public record AuthorizationRequest(
-      AuthorizationResourceType resourceType, PermissionType permissionType) {}
 
   public interface Authorizable<T extends UnifiedRecordValue> {
     AuthorizationRequest getAuthorizationRequest();
